+++ mayan/apps/document_parsing/links.py
@@ -2,7 +2,7 @@ from __future__ import unicode_literals
 
 from django.utils.translation import ugettext_lazy as _
 
-from navigation import Link
+from navigation import Link, get_cascade_condition
 
 from .icons import (
     icon_document_content, icon_document_parsing_errors_list,
@@ -44,7 +44,11 @@ link_document_type_parsing_settings = Link(
     view='document_parsing:document_type_parsing_settings',
 )
 link_document_type_submit = Link(
-    icon_class=icon_document_type_submit, text=_('Parse documents per type'),
+    condition=get_cascade_condition(
+        app_label='documents', model_name='DocumentType',
+        object_permission=permission_document_type_parsing_setup
+    ), icon_class=icon_document_type_submit,
+    text=_('Parse documents per type'),
     view='document_parsing:document_type_submit'
 )
 link_error_list = Link(