+++ mayan/apps/navigation/__init__.py
@@ -1,5 +1,6 @@
 from __future__ import unicode_literals
 
 from .classes import Link, Menu, SourceColumn  # NOQA
+from .utils import get_cascade_condition  # NOQA
 
 default_app_config = 'navigation.apps.NavigationApp'